+++ flow.js
@@ -1,7 +1,9 @@
 module.exports = {
-  "parser": "babel-eslint",
+  // warning, this is for tests
+  // you must declare this in your config as it's inherited by "extends"
+  parser: "babel-eslint",
 
-  "plugins": [
+  plugins: [
     "flow-vars",
   ],